Backport of 1.19 content to 1.18, Cleanup, Updates  - Backported 1.19 content  - Cleaned up packages  - Removed unnecessary files  - Bumped Loom to 12-Snapshot

+++ src/main/java/chronosacaria/mcdw/configs/stats/RangedWeaponStats.java
@@ -2,10 +2,10 @@ package chronosacaria.mcdw.configs.stats;
 
 public class RangedWeaponStats {
     public String material;
-    public float drawSpeed;
+    public int drawSpeed;
     public float range;
 
-    public RangedWeaponStats rangedWeaponStats (String material, float drawSpeed, float range) {
+    public RangedWeaponStats rangedWeaponStats (String material, int drawSpeed, float range) {
         this.material = material;
         this.drawSpeed = drawSpeed;
         this.range = range;
